1.增加商品分类功能 2.将分类功能接入到现有系统

+++ SystemCode/src/main/java/com/boot/security/server/convert/ExcelProduct2AiExecProduct.java
@@ -24,7 +24,7 @@ public class ExcelProduct2AiExecProduct {
         execProduct.setProductname(product.getProductName());
         execProduct.setProductcode(product.getProductCode());
         execProduct.setProductdesc(product.getProductDesc());
-        execProduct.setProducttypeid(Integer.parseInt(product.getProductTypeId()));
+        execProduct.setProducttypeid(Long.parseLong(product.getProductTypeId()));
         execProduct.setProductoriprice(new BigDecimal(product.getProductOriPrice()));
         execProduct.setProductsource(product.getProductSource());
         return execProduct;
